refactor payment models
